+++ NEWS.md
@@ -6,6 +6,7 @@
 - Added `raiser` to change the `raise` method.
 - `raise.e` and `raise.w` to define global `raise` methods
 - Added `is.problem` to check whether the object is an error or a warning
+- Fix `jab` when warning or error use a function.
 
 # erratum 2.0.0